Auto select project based on remote url in configuration page
Update release notes

+++ merge-request-integration/src/main/kotlin/net/ntworld/mergeRequestIntegration/provider/gitlab/GitlabProjectApi.kt
@@ -16,7 +16,7 @@ class GitlabProjectApi(
 ) : ProjectApi {
 
     override fun find(projectId: String): Project? {
-        val out = infrastructure.serviceBus() process GitlabFindProjectRequest(credentials, projectId)
+        val out = infrastructure.serviceBus() process GitlabFindProjectRequest(credentials, projectId.toInt())
 
         return if (out.hasError()) {
             null
